mp5/requirements.txt
numpy
matplotlib
torch
torchvision
tqdm
gradescope-utils
tensorboardX
pillow
absl-py
